split HeapDump REST vs GraphQL tests
