+++ src/test/java/itest/agent/AgentHeapDumpIT.java
@@ -16,8 +16,6 @@
 package itest.agent;
 
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
 
 import java.time.Duration;
 import java.util.List;
@@ -64,43 +62,10 @@ public class AgentHeapDumpIT extends AgentTestBase {
             }
             """;
 
-    private static final String LIST_HEAP_DUMPS_QUERY =
-            """
-            query AllTargetsHeapDumps {
-                      targetNodes {
-                        target {
-                          agent
-                          id
-                          connectUrl
-                          alias
-                          jvmId
-                          heapDumps {
-                            data {
-                              jvmId
-                              downloadUrl
-                              heapDumpId
-                              lastModified
-                              size
-                              metadata {
-                                labels {
-                                  key
-                                  value
-                                }
-                              }
-                            }
-                            aggregate {
-                              count
-                            }
-                          }
-                        }
-                      }
-                    }
-            """;
-
     @AfterEach
     void cleanupHeapDumps() {
         var variables = Map.<String, Object>of("targetIds", List.of(target.id()));
-        io.restassured.response.Response response =
+        Response response =
                 given().basePath("/")
                         .body(
                                 Map.of(
@@ -523,160 +488,6 @@ public class AgentHeapDumpIT extends AgentTestBase {
                 .body(
                         "find { it.jvmId == '" + target.jvmId() + "' }.heapDumps.size()",
                         Matchers.greaterThanOrEqualTo(1));
-    }
-
-    @Test
-    @DisabledIfEnvironmentVariable(named = "CI", matches = "true")
-    void testGraphQLListHeapDumps() throws Exception {
-        long targetId = target.id();
-
-        String jobId =
-                given().log()
-                        .all()
-                        .pathParams("targetId", targetId)
-                        .when()
-                        .post("/api/beta/diagnostics/targets/{targetId}/heapdump")
-                        .then()
-                        .log()
-                        .all()
-                        .and()
-                        .assertThat()
-                        .statusCode(200)
-                        .contentType(ContentType.TEXT)
-                        .body(Matchers.notNullValue())
-                        .and()
-                        .extract()
-                        .body()
-                        .asString();
-
-        webSocketClient.expectNotification(
-                "HeapDumpSuccess",
-                Duration.ofMinutes(2),
-                o -> jobId.equals(o.getJsonObject("message").getString("jobId")));
-
-        JsonObject uploadedNotification =
-                webSocketClient.expectNotification(
-                        "HeapDumpUploaded",
-                        Duration.ofMinutes(5),
-                        o -> target.jvmId().equals(o.getJsonObject("message").getString("jvmId")));
-
-        JsonObject message = uploadedNotification.getJsonObject("message");
-        String notificationJvmId = message.getString("jvmId");
-        JsonObject heapDumpObj = message.getJsonObject("heapDump");
-        String notificationHeapDumpId = heapDumpObj.getString("heapDumpId");
-
-        MatcherAssert.assertThat(notificationJvmId, Matchers.equalTo(target.jvmId()));
-        MatcherAssert.assertThat(notificationHeapDumpId, Matchers.notNullValue());
-
-        // Test the GraphQL Query
-        var variables = Map.<String, Object>of("targetIds", List.of(targetId));
-        Response resp =
-                given().basePath("/")
-                        .body(Map.of("query", LIST_HEAP_DUMPS_QUERY, "variables", variables))
-                        .contentType(ContentType.JSON)
-                        .log()
-                        .all()
-                        .when()
-                        .post("/api/v4/graphql")
-                        .then()
-                        .log()
-                        .all()
-                        .and()
-                        .assertThat()
-                        .statusCode(200)
-                        .extract()
-                        .response();
-
-        assertThat(resp.getStatusCode(), equalTo(200));
-        JsonObject retrievedResponse = new JsonObject(resp.asString());
-        assertThat(retrievedResponse, notNullValue());
-        assertThat(retrievedResponse.getMap().size(), not(0));
-    }
-
-    @Test
-    @DisabledIfEnvironmentVariable(named = "CI", matches = "true")
-    void testCaptureHeapDumpMutation() throws Exception {
-        JsonObject query = new JsonObject();
-        query.put(
-                "query",
-                "mutation { createHeapDump( nodes:{annotations: [\"REALM = Custom"
-                        + " Targets\"]})}");
-
-        Response response =
-                given().contentType(ContentType.JSON)
-                        .body(query.encode())
-                        .when()
-                        .post("/api/v4/graphql")
-                        .then()
-                        .statusCode(allOf(greaterThanOrEqualTo(200), lessThan(300)))
-                        .extract()
-                        .response();
-
-        webSocketClient.expectNotification("HeapDumpSuccess", Duration.ofMinutes(2));
-
-        webSocketClient.expectNotification(
-                "HeapDumpUploaded",
-                Duration.ofMinutes(5),
-                o -> target.jvmId().equals(o.getJsonObject("message").getString("jvmId")));
-
-        assertThat(response.getStatusCode(), equalTo(200));
-        assertThat(response.getBody(), notNullValue());
-    }
-
-    @Test
-    @DisabledIfEnvironmentVariable(named = "CI", matches = "true")
-    void testDeleteMutation() throws Exception {
-        JsonObject createQuery = new JsonObject();
-        createQuery.put(
-                "query",
-                "mutation { createHeapDump(nodes:{annotations:[\"REALM = Custom"
-                        + " Targets\"]})}");
-
-        // Create a Heap Dump
-        Response response =
-                given().contentType(ContentType.JSON)
-                        .body(createQuery.encode())
-                        .when()
-                        .post("/api/v4/graphql")
-                        .then()
-                        .statusCode(allOf(greaterThanOrEqualTo(200), lessThan(300)))
-                        .extract()
-                        .response();
-
-        webSocketClient.expectNotification("HeapDumpSuccess", Duration.ofMinutes(2));
-
-        assertThat(response.getStatusCode(), equalTo(200));
-        JsonObject retrieved = new JsonObject(response.body().asString());
-        assertThat(retrieved, notNullValue());
-
-        JsonObject uploadedNotification =
-                webSocketClient.expectNotification(
-                        "HeapDumpUploaded",
-                        Duration.ofMinutes(5),
-                        o -> target.jvmId().equals(o.getJsonObject("message").getString("jvmId")));
-
-        JsonObject message = uploadedNotification.getJsonObject("message");
-        String notificationHeapDumpId = message.getString("heapDumpId");
-
-        JsonObject deleteQuery = new JsonObject();
-        deleteQuery.put(
-                "query",
-                String.format(
-                        "mutation { deleteHeapDump (nodes: { annotations: [\"REALM = Custom"
-                                + " Targets\"]}, filter: { name: \"%s\"}) { name downloadUrl } }",
-                        notificationHeapDumpId));
-
-        Response deleteResponse =
-                given().contentType(ContentType.JSON)
-                        .body(deleteQuery.encode())
-                        .when()
-                        .post("/api/v4/graphql")
-                        .then()
-                        .statusCode(allOf(greaterThanOrEqualTo(200), lessThan(300)))
-                        .extract()
-                        .response();
-
-        assertThat(deleteResponse.getStatusCode(), equalTo(200));
     }
 
     protected void cleanupHeapDumpsForTarget(List<Integer> ids) {
